REF: Rename `messages` module to `events`
